% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cotan_engine.R
\name{calculate_and_store_gdi}
\alias{calculate_and_store_gdi}
\title{Calculate and Store GDI for a COTAN Object}
\usage{
calculate_and_store_gdi(
  cotan_obj,
  out_dir = ".",
  cores = 1L,
  stat_type = "S",
  rows_fraction = 0.05,
  chunk_size = 1024L
)
}
\arguments{
\item{cotan_obj}{A \code{COTAN} object.}

\item{out_dir}{Character string. Directory path where the final object will be saved. Default is \code{"."}.}

\item{cores}{Integer. Number of cores to use for parallel execution. Default is \code{1L}.}

\item{stat_type}{Character string. The type of statistic to use for GDI calculation ("S" or "G"). Default is "S".}

\item{rows_fraction}{Numeric. Fraction of rows to use for chunking in GDI. Default is 0.05.}

\item{chunk_size}{Integer. Size of the data chunks for GDI calculation. Default is 1024L.}
}
\value{
An updated \code{COTAN} object containing the calculated GDI metrics.
}
\description{
Computes the Global Differential Expression Index (GDI), stores it, and saves the object.
}
