% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cotan_engine.R
\name{calculate_coex}
\alias{calculate_coex}
\title{Calculate COEX for a COTAN Object}
\usage{
calculate_coex(
  cotan_obj,
  act_on_cells = FALSE,
  return_pp_fract = FALSE,
  optimize_for_speed = TRUE,
  device_str = "cuda"
)
}
\arguments{
\item{cotan_obj}{A \code{COTAN} object.}

\item{act_on_cells}{Boolean. Act on cells. Default is FALSE.}

\item{return_pp_fract}{Boolean. Return PP fraction. Default is FALSE.}

\item{optimize_for_speed}{Boolean. Try using the torch library to run matrix calculations. Default is TRUE.}

\item{device_str}{Character string. Device to use for calculations ("cpu", "cuda"). Default is "cuda".}
}
\value{
An updated \code{COTAN} object containing the calculated COEX matrix.
}
\description{
Calculates the genes' co-expression (COEX) matrix for a prepared COTAN object.
}
