% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cotan_engine.R
\name{clean_cotan_data}
\alias{clean_cotan_data}
\title{Clean COTAN Object}
\usage{
clean_cotan_data(
  cotan_obj,
  cells_cutoff = 0.003,
  genes_cutoff = 0.002,
  cells_threshold = 0.99,
  genes_threshold = 0.99
)
}
\arguments{
\item{cotan_obj}{A \code{COTAN} object.}

\item{cells_cutoff}{Numeric. Fraction of cells a gene must be expressed in to be kept (Default 0.003).}

\item{genes_cutoff}{Numeric. Fraction of genes a cell must express to be kept (Default 0.002).}

\item{cells_threshold}{Numeric. Genes expressed in > fraction of cells marked as fully-expressed (Default 0.99).}

\item{genes_threshold}{Numeric. Cells expressing > fraction of genes marked as fully-expressing (Default 0.99).}
}
\value{
A cleaned \code{COTAN} object with updated \code{nu} estimators.
}
\description{
Applies the basic COTAN clean() method to remove sparse genes and cells.
Mitochondrial filtering is assumed to be handled prior to this step.
}
