% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cotan_config.R
\name{config_cotan_workflow}
\alias{config_cotan_workflow}
\title{Configure the COTAN workflow with Log Rotation}
\usage{
config_cotan_workflow(
  output_dir = "./",
  log_file_name = "cotan_pipeline.log",
  parallel = TRUE,
  logging_level = 2L
)
}
\arguments{
\item{output_dir}{Character. Directory where logs and outputs will be saved.}

\item{log_file_name}{Character. Name of the log file.}

\item{parallel}{Boolean. Enable parallel processing.}

\item{logging_level}{Integer. Logging verbosity level (0-3).}
}
\value{
The normalized path to the output directory.
}
\description{
Configure the COTAN workflow with Log Rotation
}
