% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_seurat.R
\name{filter_iterative_sparsity}
\alias{filter_iterative_sparsity}
\title{Iteratively Filter Sparse Matrix in Seurat Object}
\usage{
filter_iterative_sparsity(
  seurat_obj,
  cells_cutoff = 0.003,
  genes_cutoff = 0.002
)
}
\arguments{
\item{seurat_obj}{A \code{Seurat} object.}

\item{cells_cutoff}{Numeric. Fraction of cells a feature must be expressed in. Default is 0.003.}

\item{genes_cutoff}{Numeric. Fraction of features a cell must express. Default is 0.002.}
}
\value{
A new, subsetted \code{Seurat} object with the converged dimensions.
}
\description{
Extracts the raw sparse matrix from a Seurat object and iteratively applies
cutoffs for minimum expressing cells per gene and minimum expressed genes per cell.
It loops on the dgCMatrix until convergence (zero features or cells dropped),
permanently solving the "ghost" transcript problem before COTAN conversion.
}
