% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_seurat.R
\name{filter_spliced_transcripts}
\alias{filter_spliced_transcripts}
\title{Filter Seurat Object to Retain Only Spliced Transcripts}
\usage{
filter_spliced_transcripts(seurat_obj, prefix = "^ENST")
}
\arguments{
\item{seurat_obj}{A \code{Seurat} object.}

\item{prefix}{Character string. The regex prefix to match spliced transcripts. Default is "^ENST".}
}
\value{
A new, filtered \code{Seurat} object containing only the matched features.
}
\description{
Filters the features of a Seurat object to keep only those that match a specific
prefix. This is used to retain mature, spliced transcripts (e.g., "ENST")
and discard unspliced/intronic regions (e.g., "ENSG").
}
