% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_cotan_execution}
\alias{log_cotan_execution}
\title{Handle the start and end of a function execution}
\usage{
log_cotan_execution(func_name, is_complete = FALSE, level = 2L)
}
\arguments{
\item{func_name}{Character. The name of the function.}

\item{is_complete}{Logical. If \code{TRUE}, prints a completion footer with
different characters. Default \code{FALSE}.}

\item{level}{Integer. Logging level for COTAN. Default \code{2L}.}
}
\value{
No return value.
}
\description{
Logs a formatted start or end message for a COTAN function.
}
