% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_debug}
\alias{log_debug}
\title{Log a debug message}
\usage{
log_debug(msg, level = 3L)
}
\arguments{
\item{msg}{Character. The debug message.}

\item{level}{Integer. Logging level for COTAN. Default \code{3L}.}
}
\value{
No return value.
}
\description{
Log a debug message
}
