% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_header}
\alias{log_header}
\title{Print a formatted header}
\usage{
log_header(msg, is_complete = FALSE, level = 1L)
}
\arguments{
\item{msg}{Character. The text of the header.}

\item{is_complete}{Logical. If \code{TRUE}, prints a completion footer with
different characters. Default \code{FALSE}.}

\item{level}{Integer. Logging level for COTAN. Default \code{1L}.}
}
\value{
No return value.
}
\description{
Useful for separating important sections in the log. Headers can be
start-of-section or completion messages.
}
