% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_info}
\alias{log_info}
\title{Log an information message}
\usage{
log_info(msg, level = 2L)
}
\arguments{
\item{msg}{Character. The information message.}

\item{level}{Integer. Logging level for COTAN. Default \code{2L}.}
}
\value{
No return value.
}
\description{
Log an information message
}
