% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_stat}
\alias{log_stat}
\title{Log a statistical bullet point}
\usage{
log_stat(msg, level = 2L)
}
\arguments{
\item{msg}{Character. The message to log as a bullet point.}

\item{level}{Integer. Logging level for COTAN. Default \code{2L}.}
}
\value{
No return value.
}
\description{
Log a statistical bullet point
}
