% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_warn}
\alias{log_warn}
\title{Log a warning message}
\usage{
log_warn(msg, level = 0L)
}
\arguments{
\item{msg}{Character. The warning message to log.}

\item{level}{Integer. Logging level for COTAN. Default \code{0L}.}
}
\value{
No return value.
}
\description{
Prints a formatted warning message.
}
