% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cotan_visual.R
\name{save_cotan_plots}
\alias{save_cotan_plots}
\title{Generate and Save COTAN QC and Clustering Plots}
\usage{
save_cotan_plots(
  cotan_obj,
  out_dir,
  prefix = "cotan_analysis",
  genes = NULL,
  condition = "",
  statType = "S",
  GDIThreshold = 1.43,
  GDIIn = NULL
)
}
\arguments{
\item{cotan_obj}{A processed COTAN object.}

\item{out_dir}{The directory to save the PDF plots.}

\item{prefix}{A string prefix for the file names (e.g., "sample1").}

\item{genes}{A named list of genes to label in the GDI plot. Default is NULL.}

\item{condition}{A string corresponding to the condition/sample (used for the plot title). Default is "".}

\item{statType}{Type of statistic to use for GDI ("S" or "G"). Default is "S".}

\item{GDIThreshold}{The threshold level used in the uniformity check. Default is 1.43.}

\item{GDIIn}{Pre-calculated GDI dataframe to speed up the process. Default is NULL.}
}
\description{
Generate and Save COTAN QC and Clustering Plots
}
