% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_seurat.R
\name{filter_mt_transcripts}
\alias{filter_mt_transcripts}
\title{Filter Seurat Object by Mitochondrial Transcript Percentage}
\usage{
filter_mt_transcripts(seurat_obj, mt_transcripts_list, threshold = 5)
}
\arguments{
\item{seurat_obj}{A \code{Seurat} object containing single-cell RNA-seq data.}

\item{mt_transcripts_list}{A character vector containing transcript IDs.}

\item{threshold}{Numeric. Maximum acceptable MT percentage. Default is 5.}
}
\value{
A new, filtered \code{Seurat} object.
}
\description{
Calculates the percentage of mitochondrial counts per cell using a custom
list of transcript IDs and filters out cells exceeding a specified threshold.
}
