% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cotan_init.R
\name{initialize_cotan_from_seurat}
\alias{initialize_cotan_from_seurat}
\title{Initialize COTAN Object from Seurat Object}
\usage{
initialize_cotan_from_seurat(
  seurat_obj,
  geo_id,
  seq_method,
  condition,
  override = FALSE
)
}
\arguments{
\item{seurat_obj}{A \code{Seurat} object.}

\item{geo_id}{String. Identifier for the experiment.}

\item{seq_method}{String. Sequencing method (Default "10XV1").}

\item{condition}{String. Sample condition.}

\item{override}{Boolean. Silently allows overriding data for an existing condition name. Default is FALSE.}
}
\value{
An initialized \code{COTAN} object.
}
\description{
Converts a Seurat object into a COTAN object, ensuring counts are
properly rounded to integers and metadata is transferred safely.
Is thought to be used after mithocondrial filtering and before
cleaning the COTAN object. The seurat object must have a "percent.mt"
column in its metadata.
}
